// frontend/src/types/api.d.ts
export interface components {
    schemas: {
// ...
        RefreshAccessToken: {
            /** Refresh Token */
            refresh_token: string;
            /** Access Token */
            access_token: string;
            /**
             * Token Type
             * @default bearer
             */
            token_type: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}